correct stream closing if no keep-alive activated (broken in 918)

+++ src/com/alsutton/jabber/JabberStream.java
@@ -147,7 +147,7 @@ public class JabberStream implements XMLEventListener, Runnable {
      */
     
     public void close() {
-        keepAlive.destroyTask();
+        if (keepAlive!=null) keepAlive.destroyTask();
         
         dispatcher.setJabberListener( null );
         try {
